backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-extensions==3.1.5
django-utils-six==2.0
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
oauthlib==3.2.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
